# openapi.yaml
components:
  schemas:
...
    FileUploadWithAccession:
      description: A FileUpload with its accession
      properties:
        accession:
          anyOf:
          - pattern: ^GHGAF.+
            type: string
          - type: 'null'
          description: The accession number assigned to this file.
          title: Accession
        alias:
          description: The submitted alias from the metadata (unique within the box)
          title: Alias
          type: string
        box_id:
          description: The ID of the FileUploadBox this file is associated with
          format: uuid4
          title: Box Id
          type: string
        bucket_id:
          description: The name of the bucket where the file is currently stored
          title: Bucket Id
          type: string
        decrypted_sha256:
          anyOf:
          - type: string
          - type: 'null'
          description: SHA-256 checksum of the entire unencrypted file content
          title: Decrypted Sha256
        decrypted_size:
          description: The size of the unencrypted file
          title: Decrypted Size
          type: integer
        encrypted_parts_md5:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: The MD5 checksum for each file part, in sequence
          title: Encrypted Parts Md5
        encrypted_parts_sha256:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: The SHA256 checksum for each file part, in sequence
          title: Encrypted Parts Sha256
        encrypted_size:
          description: The size of the encrypted file content. When the file is in
            the inbox, this includes the Crypt4GH envelope. After re-encryption, the
            file no longer contains an envelope, so the value is slightly smaller.
          title: Encrypted Size
          type: integer
        failure_reason:
          anyOf:
          - type: string
          - type: 'null'
          description: The reason for upload or interrogation failure, if applicable.
          title: Failure Reason
        id:
          description: Unique identifier for the file upload
          format: uuid4
          title: Id
          type: string
        object_id:
          description: The ID of the file specific to its S3 bucket
          format: uuid4
          title: Object Id
          type: string
        part_size:
          description: The number of bytes in each file part (last part is likely
            smaller)
          title: Part Size
          type: integer
        secret_id:
          anyOf:
          - type: string
          - type: 'null'
          description: The internal ID of the Data Hub-generated decryption secret
          title: Secret Id
        state:
          default: init
          description: The state of the FileUpload
          enum:
          - init
          - inbox
          - failed
          - cancelled
          - interrogated
          - awaiting_archival
          - archived
          title: State
          type: string
        state_updated:
          description: Timestamp of when state was updated
          format: date-time
          title: State Updated
          type: string
        storage_alias:
          description: The storage alias of the Data Hub housing the file
          title: Storage Alias
          type: string
      required:
      - id
      - box_id
      - alias
      - state_updated
      - storage_alias
      - bucket_id
      - object_id
      - decrypted_size
      - encrypted_size
      - part_size
      title: file_upload
      type: object
    GrantAccessRequest:
      description: Request model for granting upload access to a user.
      properties:
        box_id:
          description: ID of the upload box
          format: uuid4
          title: Box Id
          type: string
        iva_id:
          description: ID of the IVA verification
          format: uuid4
          title: Iva Id
          type: string
        user_id:
          description: ID of the user to grant access to
          format: uuid4
          title: User Id
          type: string
        valid_from:
          description: Start date of validity
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          format: date-time
          title: Valid Until
          type: string
      required:
      - box_id
      - user_id
      - iva_id
      - valid_from
      - valid_until
      title: GrantAccessRequest
      type: object
    GrantId:
      description: The ID of an access grant.
      properties:
        id:
          description: Internal grant ID (same as claim ID)
          format: uuid4
          title: Id
          type: string
      required:
      - id
      title: GrantId
      type: object
    GrantWithBoxInfo:
      description: An UploadGrant with the ResearchDataUploadBox title and description.
      properties:
        box_description:
          description: Describes the upload box in more detail
          title: Box Description
          type: string
        box_id:
          description: ID of the upload box this grant is for
          format: uuid4
          title: Box Id
          type: string
        box_title:
          description: Short meaningful name for the box
          title: Box Title
          type: string
        created:
          description: Date of creation of this grant
          format: date-time
          title: Created
          type: string
        id:
          description: Internal grant ID (same as claim ID)
          format: uuid4
          title: Id
          type: string
        iva_id:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: ID of an IVA associated with this grant
          title: Iva Id
        user_email:
          description: The email address of the user
          format: email
          title: User Email
          type: string
        user_id:
          description: Internal user ID
          format: uuid4
          title: User Id
          type: string
        user_name:
          description: Full name of the user
          title: User Name
          type: string
        user_title:
          anyOf:
          - type: string
          - type: 'null'
          description: Academic title of the user
          title: User Title
        valid_from:
          description: Start date of validity
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          format: date-time
          title: Valid Until
          type: string
      required:
      - id
      - user_id
      - box_id
      - created
      - valid_from
      - valid_until
      - user_name
      - user_email
      - box_title
      - box_description
      title: GrantWithBoxInfo
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service providing a web-accessible management and auth layer over
    upload-path file services.
  title: GHGA Registry Service
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /upload-boxes/{box_id}/uploads:
    get:
      description: List the details of all files uploads for a research data upload
        box.
      operationId: list_upload_box_files_upload_boxes__box_id__uploads_get
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid
          title: Box Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FileUploadWithAccession'
                title: Response 200 List Upload Box Files Upload Boxes  Box Id  Uploads
                  Get
                type: array
          description: File upload information successfully retrieved.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '404':
          description: Upload box not found.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List files in upload box
      tags:
      - ResearchDataUploadBoxes
  /upload-grants:
    get:
      description: Endpoint to get the list of all upload access grants. Can be filtered
        by user ID, IVA ID, and box ID. Results are sorted by validity, box ID, user
        ID, IVA ID, and grant ID.
      operationId: get_upload_access_grants_upload_grants_get
      parameters:
      - description: The internal ID of the user
        in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The internal ID of the user
          title: User Id
      - description: The ID of the IVA
        in: query
        name: iva_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The ID of the IVA
          title: Iva Id
      - description: The ID of the upload box
        in: query
        name: box_id
        required: false
        schema:
          anyOf:
          - format: uuid4
            type: string
          - type: 'null'
          description: The ID of the upload box
          title: Box Id
      - description: Whether the grant is currently valid
        in: query
        name: valid
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Whether the grant is currently valid
          title: Valid
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GrantWithBoxInfo'
                title: Response 200 Get Upload Access Grants Upload Grants Get
                type: array
          description: Upload access grants have been fetched.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get upload access grants
      tags:
      - UploadGrants
    post:
      description: Grant upload access to a user for a single research data upload
        box. Users cannot upload any files until they have been granted access to
        a box.
      operationId: grant_upload_access_upload_grants_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantAccessRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantId'
          description: Upload access granted successfully.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '422':
          description: Validation error in request body.
      security:
      - HTTPBearer: []
      summary: Grant upload access
      tags:
      - UploadGrants
  /upload-grants/{grant_id}:
    delete:
      description: Revokes an existing upload access grant.
      operationId: revoke_upload_access_grant_upload_grants__grant_id__delete
      parameters:
      - in: path
        name: grant_id
        required: true
        schema:
          format: uuid4
          title: Grant Id
          type: string
      responses:
        '204':
          description: Upload access grant has been revoked.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '404':
          description: The upload access grant was not found.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Revoke an upload access grant
      tags:
      - UploadGrants
...